requirements.txt
# Core LangChain packages
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.1
langchain-groq>=0.0.1
langchain-mistralai>=0.0.1
langchain-cohere>=0.0.1

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Rich terminal interface
rich>=13.7.0

# Interactive components
inquirer>=3.1.3

# Type hints
typing-extensions>=4.8.0

# JSON handling
jsonschema>=4.20.0

# Note: Some versions might need adjustment based on compatibility requirements
# Install with: pip install -r requirements.txt

# Optional: Development tools
# pytest>=8.0.0
# black>=24.0.0
# isort>=5.13.0
# mypy>=1.8.0